Completely redo my personal blog theme
This goes way back to basics, relying heavily upon user configuration for style.
This assumes a light background with dark foreground (typically white and black)
with a few hardcoded elements (dividing bars, hyperlinks) but is otherwise left
to the browser's defaults.

Part of the intention here is to render decently in a text-based browser like Lynx.

Mostly this is an attempt to remove needless cruft.

- Remove most graphical flourishes
- Remove the custom theme directory